+++ zoo-modules/shared-module/InputLabel.svelte
@@ -1,12 +1,12 @@
 <svelte:options tag="zoo-input-label"></svelte:options>
 {#if labeltext}
-	<div>{labeltext}</div>
+	<label>{labeltext}</label>
 {/if}
 
 <style type='text/scss'>
 	@import "variables";
 
-	div {
+	label {
 		font-size: $p1-size;
 		line-height: $p1-line-height;
 		font-weight: 800;